Refactored and moved state fundtions to its own compoenent script and set it on game master
